feat: re-structure project for better readability
