+++ src/Lottie/index.js
@@ -1,0 +1,3 @@
+import Lottie from "./Lottie";
+
+export { Lottie };
