+++ src/LottieLegacy/index.js
@@ -1,8 +1,6 @@
-import Lottie from "./Lottie";
 import { Lottie as ReactLottie, LottieWithRef as ReactLottieWithRef } from "./LottieLegacy";
 
 export {
-  Lottie,
   ReactLottie,
-  ReactLottieWithRef
+  ReactLottieWithRef,
 }
