+++ src/stories/assets/index.js
@@ -1,0 +1,9 @@
+import * as PinJump from "./pinjump.json";
+import * as BeatingHeart from "./beating-heart.json";
+import * as TwitterHeart from "./TwitterHeart.json";
+
+export {
+  PinJump,
+  BeatingHeart,
+  TwitterHeart
+}
